Adding color and shape to class Fish, handle new privileges
Can't be arsed to set color and shape a better way than in the 
constructor. A builder pattern seems overkill, but maybe that's my naive understanding of how to build things.

Why do this? I set the attribute and method privileges to be more
restrictive, thereby making it impossible to override this.color, and 
this.shape in the sub class constructor.